+++ src/utils/context/index.js
@@ -1,17 +1,8 @@
 import { createContext } from 'react';
-import { lightTheme, darkTheme } from '../../styles/layout/__themes';
-
-const setDefaultTheme = () => {
-  const today = new Date();
-  const hour = today.getHours();
-
-  if (hour >= 19 || hour <= 6) return darkTheme;
-  else if (hour > 6 && hour < 19) return lightTheme;
-  else return lightTheme;
-};
+import { lightTheme } from '../../styles/layout/__themes';
 
 export const ThemeContext = createContext({
-  currentTheme: setDefaultTheme(),
+  currentTheme: lightTheme,
   transition: '0.5s',
   toggleTheme: () => {},
 });